Do not adjust avail_min sw parameter
pcms that force a smaller buffer size (eg `dmix` with a buffer
size of 16384 frames) can underrun if dswitch waits too long
between writes. A negative side effect of this change is that it
results in higher cpu usage for some applications (eg MPD)
because of increased frequency of poll() wakeups.